sdk/identity/identity/test/public/node/azurePowerShellCredential.spec.ts
// Copyright (c) Microsoft Corporation.
// Licensed under the MIT license.

/* eslint-disable @typescript-eslint/no-non-null-asserted-optional-chain */

import assert from "assert";
import { MsalTestCleanup, msalNodeTestSetup } from "../../msalTestUtils";
import { AzurePowerShellCredential } from "../../../src";
import { Context } from "mocha";

describe("AzurePowerShellCredential", function() {
  let cleanup: MsalTestCleanup;
  beforeEach(function(this: Context) {
    cleanup = msalNodeTestSetup(this).cleanup;
  });
  afterEach(async function() {
    await cleanup();
  });

  const scope = "https://vault.azure.net/.default";

  it("authenticates", async function() {
    const credential = new AzurePowerShellCredential();

    const token = await credential.getToken(scope);
    assert.ok(token?.token);
    assert.ok(token?.expiresOnTimestamp! > Date.now());
  });
});
